Fix for switching back to original Team of enemies
The EnemyBasics.Enemies array size is only ever 62, and the Ids of the objects are not sequential in this array (There is a gap between Lamia Id: 42 and Mage Id: 128).

So this would only work for enemies with an Id of 42 or less. Better to check with Linq on the actual Id of the object.